requirements.txt
# Core app framework
gradio>=4.0.0

# PDF processing
PyPDF2>=3.0.1
PyMuPDF>=1.23.0  # High-speed PDF text extraction and rendering

# Image processing
Pillow>=10.0.0
pytesseract>=0.3.10

# Numerical & ML utilities
numpy>=1.24.0
scikit-learn>=1.3.0

# Embeddings
sentence-transformers>=2.2.2
# sentence-transformers relies on PyTorch
torch>=2.0.0

# Multimodal Vintern embeddings
transformers==4.48.0
accelerate>=0.34.0
decord>=0.6.0
# flash-attn>=2.5.8  # DISABLED: Linux/CUDA only, causes installation issues on Windows
# Note: Vintern embeddings will work without flash-attn, just slightly slower on GPU

# Google Gemini API
google-generativeai>=0.7.0

# LangChain for RAG orchestration
langchain>=0.1.0
langchain-community>=0.1.0
langchain-google-genai>=0.0.6
tiktoken>=0.5.0

# Vector stores
faiss-cpu>=1.7.4  # Local vector store (use faiss-gpu for GPU support)

FastAPI>=0.115.0
uvicorn>=0.32.0

# Authentication
authlib>=1.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
google-auth>=2.23.0

# Environment variables
python-dotenv>=1.0.0
dotenv

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
alembic>=1.13.0
pgvector>=0.3.1  # PostgreSQL vector extension for embeddings

sqlalchemy
psycopg2